+++ src/app/page.tsx
@@ -1,25 +1,9 @@
-import CheckBox from '@/components/atom/CheckBox'
-import { HOKKAIDO_TOHOKU_PREF_IDS } from '@/constant'
-import { fetchResasPrefectures } from '@/server/lib/api'
+import PrefectureChoice from '@/app/_TopPage/PrefectureChoice'
 
 export default async function TopPage() {
-  const prefectures = await fetchResasPrefectures()
-  const hokkaidoTohokuPrefs = prefectures.result.filter((v) =>
-    HOKKAIDO_TOHOKU_PREF_IDS.includes(v.prefCode),
-  )
-
   return (
-    <main className='mx-auto mt-16 max-w-screen-lg rounded-md bg-white'>
-      <div className='p-16'>
-        <div className='mb-4'>
-          <h2 className='mb-2 text-xl'>北海道・東北</h2>
-          <div className='flex gap-6'>
-            {hokkaidoTohokuPrefs.map((v) => (
-              <CheckBox key={v.prefCode} label={v.prefName} />
-            ))}
-          </div>
-        </div>
-      </div>
+    <main className='mx-auto mt-16 max-w-screen-lg rounded-md bg-white p-16'>
+      <PrefectureChoice />
     </main>
   )
 }
